Extract estimation result from doc.
