+++ src/main/java/oripa/persistent/doc/exporter/ModelExporterOBJ.java
@@ -1,38 +1,31 @@
 /**
- * ORIPA - Origami Pattern Editor
- * Copyright (C) 2005-2009 Jun Mitani http://mitani.cs.tsukuba.ac.jp/
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * ORIPA - Origami Pattern Editor Copyright (C) 2005-2009 Jun Mitani
+ * http://mitani.cs.tsukuba.ac.jp/
+ * 
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software
+ * Foundation, either version 3 of the License, or (at your option) any later
+ * version.
+ * 
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+ * details.
+ * 
+ * You should have received a copy of the GNU General Public License along with
+ * this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
 package oripa.persistent.doc.exporter;
 
-import java.io.IOException;
-
-import oripa.doc.Doc;
-import oripa.domain.fold.halfedge.OrigamiModel;
-import oripa.persistent.entity.exporter.OrigamiModelExporterOBJ;
-
-/**
- *
- *
- * @author MITANI Jun, OUCHI Koji
- *
- */
-// export folded model
-public class ModelExporterOBJ implements DocExporter {
-
+///**
+// *
+// *
+// * @author MITANI Jun, OUCHI Koji
+// *
+// */
+//// export folded model
+//public class ModelExporterOBJ implements DocExporter {
 //	public static void export_bk(final EstimationEntityHolder doc, final String filepath)
 //			throws Exception {
 //		OrigamiModel origamiModel = doc.getOrigamiModel();
@@ -71,15 +64,14 @@ public class ModelExporterOBJ implements DocExporter {
 //		bw.close();
 //	}
 
-	@Override
-	public boolean export(final Doc doc, final String filePath)
-			throws IOException, IllegalArgumentException {
-		OrigamiModelExporterOBJ exporter = new OrigamiModelExporterOBJ();
-
-		OrigamiModel origamiModel = doc.getOrigamiModel();
-
-		return exporter.export(origamiModel, filePath);
-
-	}
-
-}
+//	@Override
+//	public boolean export(final Doc doc, final String filePath)
+//			throws IOException, IllegalArgumentException {
+//		OrigamiModelExporterOBJ exporter = new OrigamiModelExporterOBJ();
+//
+//		OrigamiModel origamiModel = doc.getOrigamiModel();
+//
+//		return exporter.export(origamiModel, filePath);
+//
+//	}
+//}
